+++ packages/react/src/components/CardCodeEditor/CardCodeEditor.jsx
@@ -10,7 +10,7 @@ import Button from '../Button';
 import deprecate from '../../internal/deprecate';
 import CodeEditor from '../CodeEditor/CodeEditor';
 
-const { iotPrefix } = settings;
+const { iotPrefix, prefix } = settings;
 
 const propTypes = {
   ...ComposedModal.propTypes, // eslint-disable-line react/forbid-foreign-prop-types
@@ -96,8 +96,8 @@ const CardCodeEditor = ({
 
   return (
     <ComposedModal
-      // TODO: remove deprecated testID in v3 and pass testId to override defaults
-      // testID={`${testID || testId}-modal`}
+      data-testid={`${testId}--card-code-editor-modal`}
+      selectorPrimaryFocus={`.${prefix}--modal-header`}
       className={classnames(`${iotPrefix}--editor`, {
         [`${iotPrefix}--editor__expanded`]: isExpanded,
       })}