+++ packages/react/src/components/List/HierarchyList/HierarchyListReorderModal/HierarchyListReorderModal.jsx
@@ -8,7 +8,7 @@ import { settings } from '../../../../constants/Settings';
 
 import RadioButtonGroup from './HierarchyReorderModalRadioGroup';
 
-const { iotPrefix } = settings;
+const { iotPrefix, prefix } = settings;
 
 const propTypes = {
   /** ListItems to be displayed */
@@ -168,6 +168,7 @@ const HierarchyListReorderModal = ({
 
   return (
     <ComposedModal
+      selectorPrimaryFocus={`.${prefix}--modal-header`}
       className={`${iotPrefix}--hierarchy-list-bulk-modal`}
       open={open}
       header={{